refactor: moved pet document filtering logic to the use case instead of the route loader
