+++ react-client/src/domain/repository/pet/GetPetDocumentsRepository.ts
@@ -1,5 +1,6 @@
+import { PetDocumentTypeId } from "~/domain/models/pet/PetDocument";
 import { PetRecord } from "~/domain/models/pet/PetRecords";
 
 export interface GetPetDocumentsRepository {
-  query(petId: string): Promise<PetRecord[]>;
+  query(petId: string, type: PetDocumentTypeId): Promise<PetRecord[]>;
 }
