+++ react-client/src/routes/my-pets/petId/documents/documentType/useDocumentTypeIndexViewModel.tsx
@@ -1,13 +1,11 @@
-import { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { LoaderFunction, useLoaderData } from "react-router-typesafe";
-import { PetRecord } from "~/components/Pet/types/PetRecordsTypes";
 
+import { isValidPetDocumentId } from "~/domain/models/pet/PetDocument";
+import { GetPetDocumentsUseCase } from "~/domain/useCases/pet/GetPetDocumentsUseCase";
+import { requireAuthToken } from "~/util/authUtil";
 import { invariant, invariantResponse } from "~/util/invariant";
 import { usePetProfileContext } from "../../usePetProfileLayoutViewModel";
-import { isValidPetDocumentId } from "../petDocumentTypeUtils";
-import { requireAuthToken } from "~/util/authUtil";
-import { GetPetDocumentsUseCase } from "~/domain/useCases/pet/GetPetDocumentsUseCase";
 
 export const loader = (async ({ params }) => {
   const { petId, documentType } = params;
@@ -20,38 +18,19 @@ export const loader = (async ({ params }) => {
   const authToken = requireAuthToken();
   const useCase = new GetPetDocumentsUseCase(authToken);
 
-  try {
-    const allDocuments = await useCase.query(petId);
-    const filteredDocuments = allDocuments.filter(
-      (doc) => doc.recordType?.toLowerCase() === documentType
-    );
+  const documents = await useCase.query(petId, documentType);
 
-    return {
-      id: documentType,
-      petId,
-      documents: filteredDocuments,
-    };
-  } catch (error) {
-    console.error("Failed to fetch pet records:", error);
-    // You might want to throw an error here or return a specific error state
-    return {
-      id: documentType,
-      petId,
-      documents: [],
-      error: "Failed to fetch documents",
-    };
-  }
+  return {
+    id: documentType,
+    petId,
+    documents,
+  };
 }) satisfies LoaderFunction;
 
 export const useDocumentTypeIndexViewModel = () => {
-  const {
-    id,
-    documents: initialDocuments,
-    error,
-  } = useLoaderData<typeof loader>();
+  const { id, documents } = useLoaderData<typeof loader>();
   const { documentTypes } = usePetProfileContext();
   const navigate = useNavigate();
-  const [documents, setDocuments] = useState<PetRecord[]>(initialDocuments);
 
   const documentType = documentTypes.find((dt) => dt.id === id);
   invariant(documentType, "Document type must be found here");
@@ -61,7 +40,8 @@ export const useDocumentTypeIndexViewModel = () => {
   };
 
   const onDelete = (recordId: string) => {
-    setDocuments((prev) => prev.filter((doc) => doc.id !== recordId));
+    // TODO: Implement real delete action when backend is ready
+    console.log("recordId", recordId);
   };
 
   return {
@@ -69,6 +49,5 @@ export const useDocumentTypeIndexViewModel = () => {
     documentType,
     onClose,
     onDelete,
-    error,
   };
 };
